Shrink Image Files with TinyPNG

Shrink an image's (PNG or JPG) filesize with the TinyPNG API.

tinify(
  file,
  overwrite = FALSE,
  return_path = FALSE,
  details = FALSE,
  key = NULL
)

Arguments

file
A string detailing the path to the file you wish to shrink, relative to the current working directory. Can include sub-directories and must include the file extension (.png or .jpg/.jpeg only).

overwrite
Boolean. By default, tinify will create a new file with the suffix '_tiny' and preserve the original file. Set TRUE to instead overwrite the original file, with the same filename.

return_path
Boolean. If TRUE, tinify returns the full absolute filepath to the newly shrunk file as a string.

details
Boolean. If TRUE, provides details on the amount of shrinkage (% and Kb), and the number of TinyPNG API calls made this month.

key
Optional. A string containing your TinyPNG API key. Not required if the API key is set using tinify_api(). If an API key is provided with tinify_api(), any other key provided in the function call will override the key set by tinify_api().

Value

If return_path = TRUE, a string with the full absolute path to the newly shrunk image file.

Details

You can get a TinyPNG API key from https://tinypng.com/developers.

TinyPNG is smart enough to know when you are uploading the same file again, and so will not count repeat calls of tinify() on the same image file against your monthly API usage limit. This can be useful if, for example, you are using tinify() in an RMarkdown document as it won't count against your API usage every time you knit your document.

See also

Examples

if (FALSE) {
# Shrink a PNG file

img <- system.file("extdata", "example.png", package = "tinieR")

tinify(img)

# Also works with JPEG/JPG files

img_jpg <- system.file("extdata", "example.jpg", package = "tinieR")

tinify(img_jpg)

# Return absolute path to newly shrunk file:

shrunk_img <- tinify(img, return_path = TRUE)

# Show details:

tinify(img, details = TRUE)

# Overwrite original file in place:

tinify(img, overwrite = TRUE)

# Overwrite a global API key set in tinify_api():

tinify(img, key = "NEW-API-KEY-HERE")

# You can combine any number of the above:

tinify(img,
       overwrite = TRUE,
       details = TRUE,
       return_path = TRUE)

# Plays nice with the pipe:

img %>% tinify()

# And with purrr::map for multiple files:

imgs <- c("example.png", "example2.png")

purrr::map(imgs, ~tinify(.x))

# An example method for shrinking an entire directory:

imgs_dir <- fs::dir_ls("imgs", glob = "*.png")

purrr::map(imgs_dir, ~tinify(.x, overwrite = TRUE))
}